requirements.txt
streamlit
torch
torchvision
timm
Pillow
numpy